Fix runtime resolution for latest jade
index.js used to be in the base dir, now has been moved to lib. This
solution should be resilient to any location of jade’s entry file
(main) as long as runtime.js stays in the base dir. Should be safe, as
moving it should require a major version bump by the jade project.